create wendy-CLI ; restruct wedy functions
