--- installer/templates/engine.sh.j2
@@ -1,4 +1,0 @@
-
-export BOB_HOME_DIRECTORY="{{ BOB_HOME_DIRECTORY }}"
-export BOB_LISTEN_ADDRESS="{{ BOB_LISTEN_ADDRESS }}"
-export BOB_HTML_DIRECTORY="{{ BOB_HTML_DIRECTORY }}"
