+++ installer/templates/engine.yaml.j2
@@ -1,11 +1,11 @@
 ---
 
-bob_listen_address: "{{ BOB_LISTEN_ADDRESS }}"
-bob_html_directory: "{{ BOB_HTML_DIRECTORY }}"
+bob_home_directory: {{ BOB_HOME_DIRECTORY }}
+bob_html_directory: {{ BOB_HTML_DIRECTORY }}
+bob_listen_address: {{ BOB_LISTEN_ADDRESS }}
 bob_netboot_path: "rescue/{{ RESCUE_VERSION }}.{{ RESCUE_MINOR_VER }}"
 builder_ssh_key: "{{ BUILDER_SSH_KEY }}"
 dns_domain: "{{ BOB_DNS_DOMAIN }}"
-
 
 alpine_base_url: "{{ ALPINE_SITE }}/v{{ RESCUE_VERSION }}"
