+++ opt_bob/wendy/api_pages.py
@@ -1,0 +1,4 @@
+from flask import Flask, request
+
+from config import *
+from wendy import app
