+++ opt_bob/wendy/templates/hosts.htm
@@ -15,7 +15,7 @@
     {% for host in ctx.hosts %}
     <tr>
         <td>{{host.hostname}}</td>
-        <td>{{host.ip_addr}}</td>
+        <td>{{host.ip}}</td>
         <td>{{host.mac}}</td>
         <td>{{host.os}}</td>
         <td>{{host.version}}</td>
